requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
sqlalchemy==2.0.39
pymysql==1.1.1
numpy==2.2.4
Pillow==11.1.0
face_recognition==1.3.0
git+https://github.com/ageitgey/face_recognition_models
uuid==1.30
python-dotenv==1.0.1
python-multipart==0.0.20
setuptools==76.0.0